requirements.txt
jinja2==3.0
pygments==2.4.1
traitlets==5.6.0
ipython==7.23.1
pyspark==3.3.1
delta-spark==2.2.0

